Get things to compile after refactor
