+++ Cargo.toml
@@ -25,6 +25,7 @@ ark-bls12-377 = "0.4"
 ark-ed-on-bls12-377 = { version = "0.4", features = ["r1cs"] }
 ark-groth16 = { version = "0.4", default-features=false, optional=true }
 ark-snark = { version = "0.4", optional=true }
+cfg-if = "1.0"
 zeroize = {version ="1.7", default-features=false}
 hashbrown = "0.14.3"
