+++ proptest-regressions/fields/fp/arkworks.txt
@@ -1,0 +1,7 @@
+# Seeds for failure cases proptest has generated in the past. It is
+# automatically read and these particular cases re-run before any
+# novel cases are generated.
+#
+# It is recommended to check this file in to source control so that
+# everyone who runs the test benefits from these saved cases.
+cc cdd619576f50197f2531266a43db3be354532b33cd74b87be7ca92df4a5c0715 # shrinks to a = Fp(0x000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000)
