+++ src/fields/fp/u32.rs
@@ -1,2 +1,5 @@
-// pub mod fiat;
-// pub mod wrapper;
+pub(crate) mod arkworks_constants;
+pub mod fiat;
+pub mod wrapper;
+
+pub use wrapper::Fp;
