+++ src/fields/fp/u64.rs
@@ -1,2 +1,5 @@
-pub mod fiat;
+pub(crate) mod arkworks_constants;
+pub(crate) mod fiat;
 pub mod wrapper;
+
+pub use wrapper::Fp;
