+++ src/fields/fp/u64/wrapper.rs
@@ -1,37 +1,62 @@
 use super::fiat;
 
 #[derive(Copy, Clone)]
-pub struct Fp(pub fiat::FpMontgomeryDomainFieldElement);
+pub struct Fp(pub(crate) fiat::FpMontgomeryDomainFieldElement);
+
+impl PartialEq for Fp {
+    fn eq(&self, other: &Self) -> bool {
+        let sub = self.sub(other);
+        let mut check_word = 0;
+        fiat::fp_nonzero(&mut check_word, &sub.0 .0);
+        check_word == 0
+    }
+}
+
+impl Eq for Fp {}
 
 impl Fp {
-    pub fn from_bytes(bytes: &[u8; 48]) -> Self {
-        let mut x_non_montgomery = fiat::FpNonMontgomeryDomainFieldElement([0; 6]);
+    pub fn from_le_limbs(limbs: [u64; 6]) -> Self {
+        let x_non_monty = fiat::FpNonMontgomeryDomainFieldElement(limbs);
         let mut x = fiat::FpMontgomeryDomainFieldElement([0; 6]);
-
-        fiat::fp_from_bytes(&mut x_non_montgomery.0, &bytes);
-        fiat::fp_to_montgomery(&mut x, &x_non_montgomery);
-
+        fiat::fp_to_montgomery(&mut x, &x_non_monty);
         Self(x)
     }
 
-    pub fn to_bytes(&self) -> [u8; 48] {
+    pub fn from_bytes(bytes: &[u8; 48]) -> Self {
         let mut x_non_montgomery = fiat::FpNonMontgomeryDomainFieldElement([0; 6]);
-        let mut bytes = [0u8; 48];
+        fiat::fp_from_bytes(&mut x_non_montgomery.0, &bytes);
+        Self::from_le_limbs(x_non_montgomery.0)
+    }
 
+    pub fn to_le_limbs(&self) -> [u64; 6] {
+        let mut x_non_montgomery = fiat::FpNonMontgomeryDomainFieldElement([0; 6]);
         fiat::fp_from_montgomery(&mut x_non_montgomery, &self.0);
-        fiat::fp_to_bytes(&mut bytes, &x_non_montgomery.0);
+        x_non_montgomery.0
+    }
 
+    pub fn to_bytes_le(&self) -> [u8; 48] {
+        let mut bytes = [0u8; 48];
+        fiat::fp_to_bytes(&mut bytes, &self.to_le_limbs());
         bytes
     }
 
-    pub fn zero() -> Self {
+    pub const fn from_montgomery_limbs(limbs: [u64; 6]) -> Self {
+        Self(fiat::FpMontgomeryDomainFieldElement(limbs))
+    }
+
+    pub const fn zero() -> Self {
         Self(fiat::FpMontgomeryDomainFieldElement([0; 6]))
     }
 
-    pub fn one() -> Self {
-        let mut one = Self::zero();
-        fiat::fp_set_one(&mut one.0);
-        one
+    pub const fn one() -> Self {
+        Self(fiat::FpMontgomeryDomainFieldElement([
+            202099033278250856,
+            5854854902718660529,
+            11492539364873682930,
+            8885205928937022213,
+            5545221690922665192,
+            39800542322357402,
+        ]))
     }
 
     pub fn square(&self) -> Self {
@@ -119,25 +144,25 @@ impl Fp {
         Some(Fp(result))
     }
 
-    pub fn add(self, other: Fp) -> Fp {
+    pub fn add(self, other: &Self) -> Self {
         let mut result = fiat::FpMontgomeryDomainFieldElement([0; 6]);
         fiat::fp_add(&mut result, &self.0, &other.0);
         Fp(result)
     }
 
-    pub fn sub(self, other: Fp) -> Fp {
+    pub fn sub(self, other: &Self) -> Self {
         let mut result = fiat::FpMontgomeryDomainFieldElement([0; 6]);
         fiat::fp_sub(&mut result, &self.0, &other.0);
         Fp(result)
     }
 
-    pub fn mul(self, other: Fp) -> Fp {
+    pub fn mul(self, other: &Self) -> Self {
         let mut result = fiat::FpMontgomeryDomainFieldElement([0; 6]);
         fiat::fp_mul(&mut result, &self.0, &other.0);
         Fp(result)
     }
 
-    pub fn neg(self) -> Fp {
+    pub fn neg(self) -> Self {
         let mut result = fiat::FpMontgomeryDomainFieldElement([0; 6]);
         fiat::fp_opp(&mut result, &self.0);
         Fp(result)
